+++ nest-forms-backend/src/nases/nases.service.ts
@@ -30,6 +30,7 @@ import BaConfigService from '../config/ba-config.service'
 import ConvertPdfService from '../convert-pdf/convert-pdf.service'
 import { FormFilesReadyResultDto } from '../files/files.dto'
 import FilesService from '../files/files.service'
+import { RabbitPayloadDto } from '../form-delivery-consumer/form-delivery-consumer.dto'
 import FormValidatorRegistryService from '../form-validator-registry/form-validator-registry.service'
 import { FormUpdateBodyDto } from '../forms/dtos/forms.requests.dto'
 import {
@@ -38,7 +39,6 @@ import {
 } from '../forms/forms.errors.enum'
 import FormsService from '../forms/forms.service'
 import { getUserFormFields } from '../forms-v2/utils/get-user-form-fields'
-import { RabbitPayloadDto } from '../nases-consumer/nases-consumer.dto'
 import PrismaService from '../prisma/prisma.service'
 import RabbitmqClientService from '../rabbitmq-client/rabbitmq-client.service'
 import { ErrorsEnum } from '../utils/global-enums/errors.enum'
