+++ nest-forms-backend/src/utils/guards/dtos/error.dto.ts
@@ -1,10 +1,10 @@
 import { ConvertErrorsEnum } from '../../../convert/errors/convert.errors.enum'
 import { FilesErrorsEnum } from '../../../files/files.errors.enum'
+import { EmailFormsErrorsEnum } from '../../../form-delivery-consumer/subservices/dtos/email-forms.errors.enum'
+import { WebhookErrorsEnum } from '../../../form-delivery-consumer/subservices/dtos/webhook.errors.enum'
 import { FormsErrorsEnum } from '../../../forms/forms.errors.enum'
 import { GinisTaskErrorEnum } from '../../../ginis/errors/ginis-tasks.errors.enum'
 import { NasesErrorsEnum } from '../../../nases/nases.errors.enum'
-import { EmailFormsErrorsEnum } from '../../../nases-consumer/subservices/dtos/email-forms.errors.enum'
-import { WebhookErrorsEnum } from '../../../nases-consumer/subservices/dtos/webhook.errors.enum'
 import { ScannerClientErrorsEnum } from '../../../scanner-client/scanner-client.errors.enum'
 import { SignerErrorsEnum } from '../../../signer/signer.errors.enum'
 import { StatusErrorsEnum } from '../../../status/errors/status.errors.enum'
